+++ Project_2-master/app/Console/Commands/ScrapeGoogleScholar.php
@@ -34,8 +34,9 @@ class ScrapeGoogleScholar extends Command
     {
         $scholarId = $this->argument('scholar_id');
         $author = new PublicationRetrieval();
-        $author->getAuthor("cYJ4r_BHQR8C");
-        $data = $author->getPaper("Science and technology for water purification in the coming decades","https://scholar.google.com/citations?view_op=view_citation&hl=en&user=7muexxwAAAAJ&citation_for_view=7muexxwAAAAJ:u5HHmVD_uO8C");
+        //$author->getAuthor("cYJ4r_BHQR8C");
+        $data = $author->getPaperOpenAlxe("Service priority classification using machine learning");
+        dd($data);
         return 0;