+++ Project_2-master/app/Http/Controllers/UpDateUserPaperController.php
@@ -23,6 +23,7 @@ class UpDateUserPaperController extends Controller
             echo implode(", ",$PaperAuthorFName) . "\n";
             echo "**************\n";    
             $updateUserPaper = $ServiceUserPaper->assignAuthorsToPaper($paper->id,$PaperAuthorFName);
+            echo "**************\n";    
         }
     }
 }