+++ Project_2-master/resources/views/dashboards/users/layouts/user-dash-layout.blade.php
@@ -208,14 +208,14 @@ scratch. This page gets rid of all links and provides the needed markup only.
                         </div>
                     </li>
                     @endcan
-                    @can('export')
+                    {{-- @can('export')
                     <li class="nav-item">
                         <a class="nav-link" href="{{route('exportfile')}}" >
                             <i class="menu-icon mdi mdi-file-export"></i>
                             <span class="menu-title">Export</span>
                         </a>
                     </li>
-                    @endcan
+                    @endcan --}}
                     @can('user-list')
                     <li class="nav-item nav-category">Admin</li>
                     <li class="nav-item">