+++ changeLog.md
@@ -5,6 +5,8 @@
 - เพิ่ม class GetDataReport เพื่อ query ข้อมูลออกไปปริน
 - ใน class GetDataReport เพิ่ม function ส่งออกข้อมูลไปเพื่อไปปริ้น โดยมี fucntion ส่งออกดั้งนี้ getAuthorData, getPaperData, getOtherWorkData,getBookData
 - เพิ่ม document ของวิธีใช้ class GetDataReport
+- เพิ่ม pdf template เพื่อจัด format นำไปใช้กับฟังก์ชันสร้าง pdf
+- ฟังก์ชันสร้าง pdf และสามารถดาวน์โหลดได้
 ### Update
 - ปรับ function getPaperOpenAlxe ใน class PublicationRetrieval โดยแยกส่วนหีวออกมาสร้าง functionใหม่่ คือ getDataOpenAlex
 ### Bug fix